Add support to the `make` module for targets with options
In order to run a `make` target with options, we need to not
assume that the target that is passed in to the module is to
be interpreted as one command-line argument. This allows us
to support targets with options like:
  $ make some-target -o some-file

Where the `some-target` target skips rebuilding `some-file`
even if the file is old. Currently, the implementation would
attempt to run the `some-target -o some-file` target, which
does not exist.

`make` does not allow for targets with spaces in their names,
so we should be safe to break up the command-line arguments
like this and not risk mutilating a valid target name.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>